Add files via upload
My hometown, Kushtia, is a beautiful place in Bangladesh filled with history, culture, and natural beauty. It is home to the famous Shilaidaha Kuthibari of Rabindranath Tagore and the shrine of the spiritual Baul singer Lalon Shah. Visitors can enjoy the scenic Padma River, old mosques, and peaceful countryside. A trip to Kushtia offers a unique experience of music, heritage, and tradition.